extract high_precision math into separate file and clean up (and remove) tests
