+++ src/ema/mod.rs
@@ -1,6 +1,8 @@
 mod math;
 
 #[cfg(test)]
+pub(crate) mod high_precision;
+#[cfg(test)]
 mod invariants;
 #[cfg(test)]
 mod tests;
